+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -4,9 +4,9 @@ import android.app.Application
 import woowacourse.shopping.data.cartRepository.CartDatabase
 import woowacourse.shopping.data.cartRepository.CartRepository
 import woowacourse.shopping.data.cartRepository.CartRepositoryImpl
-import woowacourse.shopping.data.productsRepository.DummyProducts
 import woowacourse.shopping.data.productsRepository.ProductRepository
 import woowacourse.shopping.data.productsRepository.ProductRepositoryImpl
+import woowacourse.shopping.data.storage.DummyProducts
 
 class ShoppingApplication : Application() {
     val cartDatabase: CartDatabase by lazy {
@@ -14,7 +14,7 @@ class ShoppingApplication : Application() {
     }
     val productRepository: ProductRepository by lazy {
         ProductRepositoryImpl.initialize(
-            DummyProducts.value,
+            storage = DummyProducts,
         )
     }
     val cartRepository: CartRepository by lazy { CartRepositoryImpl.initialize(cartDatabase) }
