+++ app/src/main/java/woowacourse/shopping/data/storage/ProductStorage.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.data.storage
+
+import woowacourse.shopping.domain.Product
+
+interface ProductStorage {
+    fun getProducts(): List<Product>
+
+    fun getProducts(limit: Int): List<Product>
+}
